matlab.el (matlab-eval): Proper sync eval, at long last
* matlab.el (matlab-bg-eval-buffer): New defvar.
(matlab-eval-filter): Use `matlab-bg-eval-buffer'.
(matlab-eval): Update.

** Using a background buffer for eval:

1. Get the process of the shell buffer.
2. Assign an empty `matlab-bg-eval-buffer' to the process (replacing the
shell buffer that we don't want to change).
3. `process-send-string'
4. Collect the output.
5. Reset the process buffer to the shell buffer.

This fixes the problem of using "C-M-i" in the shell buffer, besides
making the code cleaner.

** Implementing proper sync eval:

The problem with `process-sentinel' is that it's fed the output of the
process in chunks. We never really know when we are on the last chunk.

A heuristic is used:
1. `matlab-eval' sets `matlab-eval-done' to nil and calls
`process-send-string'.
2. `matlab-eval-filter' checks if the chunk ends in ">> ", the MATLAB
prompt. If so, that means that was the last chunk and set
`matlab-eval-done' to t.
3. `matlab-eval' starts a loop checking `matlab-eval-done' and calling
`accept-process-output' otherwise. I'm not sure if that function does
anything at all, but even if it's a noop it still works out in the end.

** Collecting the output

The original command is echoed back by MATLAB. So the output is split by
"^>>" and tested to not match command.

Additionally, there's an attempt to check if there's a "x = ..." in the
output and capture that instead.