Don't lock up on MATLAB startup
* matlab.el (matlab-middleware-loaded-p): New defvar.
(matlab-shell-mode): Instead of using `matlab-eval' to add the
middleware to path, set the variable to show that the middleware is not
yet on path. While the call to `make-comint' in `matlab-shell' is async
and doesn't lock up Emacs, `matlab-eval' is sync and will not return
until MATLAB has booted and returned a result.
(matlab-emacsdocomplete): Extract defun. Load middleware if needed.
(matlab-shell-completion-list): Update.